# contracts/center-api.openapi.yaml
openapi: 3.1.0
info:
  title: Hermes Monitoring Center API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
  /projects/{projectId}/instances/{instanceId}/logs:
    get:
      operationId: listLogs
      parameters:
        - {$ref: '#/components/parameters/ProjectId'}
        - {$ref: '#/components/parameters/InstanceId'}
      responses:
        '200':
          description: Configured log files (administrator or assigned operator only)
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/LogSource'}}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      operationId: createLog
      parameters:
        - {$ref: '#/components/parameters/ProjectId'}
        - {$ref: '#/components/parameters/InstanceId'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LogWrite'}
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [log_id]
                properties:
                  log_id: {type: integer, format: int64}
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /projects/{projectId}/instances/{instanceId}/logs/{logId}:
    put:
      operationId: updateLog
      parameters:
        - {$ref: '#/components/parameters/ProjectId'}
        - {$ref: '#/components/parameters/InstanceId'}
        - {in: path, name: logId, required: true, schema: {type: integer, format: int64}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LogWrite'}
      responses:
        '200': {description: Updated}
        '400': {$ref: '#/components/responses/BadRequest'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /projects/{projectId}/instances/{instanceId}/logs/{logId}/read:
    get:
      operationId: readLog
      parameters:
        - {$ref: '#/components/parameters/ProjectId'}
        - {$ref: '#/components/parameters/InstanceId'}
        - {in: path, name: logId, required: true, schema: {type: integer, format: int64}}
        - {in: query, name: cursor, schema: {type: string, maxLength: 100}}
      responses:
        '200':
          description: Bounded raw text; on-demand shared polling, never persisted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LogRead'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /projects/{projectId}/instances/{instanceId}:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/InstanceId'}
    put:
      operationId: updateInstance
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/InstanceUpdate'}
      responses:
        '200': {description: Instance updated}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
    delete:
      operationId: disableInstance
      responses:
        '204': {description: Instance disabled}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /projects/{projectId}/instances/{instanceId}/test:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/InstanceId'}
    post:
      operationId: testInstance
      responses:
        '200':
          description: Connection result
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ActionStatus'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /projects/{projectId}/instances/{instanceId}/internal-checks:
    parameters:
      - {$ref: '#/components/parameters/ProjectId'}
      - {$ref: '#/components/parameters/InstanceId'}
    get:
      operationId: listInternalChecks
      parameters:
        - {name: monitoringEnabled, in: query, schema: {type: boolean}, description: Filter used or unused checks.}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
        - {$ref: '#/components/parameters/Sort'}
        - {$ref: '#/components/parameters/StatusFilter'}
      responses:
        '200':
          description: Internal check page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CheckPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api-checks:
    get:
      operationId: listApiChecks
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
        - {$ref: '#/components/parameters/Sort'}
        - {$ref: '#/components/parameters/StatusFilter'}
        - {name: project, in: query, schema: {$ref: '#/components/schemas/Identifier'}}
        - {name: instance, in: query, schema: {$ref: '#/components/schemas/Identifier'}}
        - {name: direction, in: query, schema: {type: string, enum: [INTERNAL, EXTERNAL]}}
        - {name: name, in: query, schema: {type: string, maxLength: 120}}
        - {name: monitoringEnabled, in: query, schema: {type: boolean}, description: Filter used or unused API checks. Omit to include both.}
      responses:
        '200':
          description: API check page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CheckPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /certificates/{id}/check:
    parameters:
      - {$ref: '#/components/parameters/CertificateId'}
    post:
      operationId: checkCertificate
      responses:
        '202':
          description: Certificate check accepted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ActionStatus'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/RateLimited'}
...
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: JSESSIONID}
  parameters:
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 0, default: 0}
    Size:
      name: size
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
    Sort:
      name: sort
      in: query
      description: Allowlisted field and direction, for example project_id,asc
      schema: {type: string}
    StatusFilter:
      name: status
      in: query
      schema: {$ref: '#/components/schemas/Status'}
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Identifier'}
    InstanceId:
      name: instanceId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Identifier'}
...
    CertificateId:
      name: id
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Forbidden:
      description: CSRF token missing or access denied
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    RateLimited:
      description: Check already running or rate limited
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
    LogWrite:
      type: object
      additionalProperties: false
      required: [name, path, encoding, enabled, pollIntervalSeconds]
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        path: {type: string, minLength: 1, maxLength: 1000}
        encoding: {type: string, enum: [UTF-8, MS949, EUC-KR]}
        enabled: {type: boolean}
        pollIntervalSeconds: {type: integer, minimum: 5, maximum: 1800, default: 10}
    LogSource:
      type: object
      required: [log_id, name, path, encoding, enabled, poll_interval_seconds]
      properties:
        log_id: {type: integer, format: int64}
        name: {type: string}
        path: {type: string}
        encoding: {type: string, enum: [UTF-8, MS949, EUC-KR]}
        enabled: {type: boolean}
        poll_interval_seconds: {type: integer}
    LogRead:
      type: object
      required: [text, cursor, reset, limited, code, retry_after_seconds]
      properties:
        text: {type: string, maxLength: 131072}
        cursor: {type: string}
        reset: {type: boolean}
        limited: {type: boolean}
        code: {type: [string, 'null']}
        retry_after_seconds: {type: integer}
    Identifier:
      type: string
      pattern: '^[a-z0-9][a-z0-9._-]{1,63}$'
    Status:
      type: string
      enum: [UP, WARN, DOWN, UNKNOWN]
    PageMetadata:
      type: object
      required: [page, size, total]
      properties:
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1, maximum: 200}
        total: {type: integer, format: int64, minimum: 0}
...
    Error:
      type: object
      required: [code]
      properties:
        code: {type: string}
    ActionStatus:
      type: object
      required: [status]
      properties:
        status: {type: string}
        code:
          type: string
          enum: [ADDRESS_NOT_ALLOWED, ALLOWLIST_CONFIG_ERROR, AUTH_ERROR, REDIRECT_REJECTED, AGENT_TIMEOUT,
            AGENT_HTTP_ERROR, IDENTITY_MISMATCH, TLS_ERROR, DNS_ERROR, CONNECTION_ERROR]
        days_remaining: {type: integer}
...
    InstanceUpdate:
      type: object
      required: [instanceId, displayName, agentBaseUrl, apiChecksEnabled, pollIntervalSeconds, enabled]
      properties:
        instanceId: {$ref: '#/components/schemas/Identifier'}
        displayName: {type: string, minLength: 1, maxLength: 120, pattern: '^\S(?:[\s\S]*\S)?$'}
        environment: {type: [string, 'null'], maxLength: 40}
        agentBaseUrl: {type: string, format: uri, minLength: 1, maxLength: 500, description: "Normalized HTTP(S) URL without credentials, query, or fragment"}
        token: {type: [string, 'null'], writeOnly: true, x-minUtf8BytesWhenNonBlank: 32, description: "Null or blank preserves the existing token; otherwise at least 32 bytes when UTF-8 encoded"}
        apiChecksEnabled: {type: boolean}
        pollIntervalSeconds: {type: integer, minimum: 15, maximum: 715827882}
        enabled: {type: boolean}
...
    ServiceProfile:
      type: object
      additionalProperties: false
      required: [title, dashboard, fields]
      properties:
        title: {type: string, minLength: 1, maxLength: 120}
        description: {type: string, maxLength: 500}
        dashboard: {type: boolean}
        fields: {type: array, minItems: 1, maxItems: 16, items: {$ref: '#/components/schemas/ServiceField'}}
    ServiceField:
      type: object
      additionalProperties: false
      required: [key, label, source, path, kind]
      properties:
        key: {type: string, pattern: '^[a-z][a-z0-9_]{0,63}$'}
        label: {type: string, minLength: 1, maxLength: 120}
        source: {type: string, enum: [DETAILS, RESPONSE_JSON]}
        path: {type: string, maxLength: 256, description: 'Simple JSON path, for example $.balance or $.data.remaining[0]. Only named fields and array indexes are supported.'}
        kind: {type: string, enum: [NUMBER, TEXT, BOOLEAN]}
        unit: {type: string, maxLength: 16}
        direction: {type: string, enum: [LOW, HIGH], default: LOW}
        required: {type: boolean, description: Missing this field prevents a healthy service information status.}
        warning: {type: [number, 'null']}
        critical: {type: [number, 'null']}
        valueLabels: {type: object, maxProperties: 10, additionalProperties: {type: string, maxLength: 120}}
        when:
          type: object
          additionalProperties: false
          required: [path, values]
          properties:
            path: {type: string, maxLength: 256}
            values:
              type: array
              minItems: 1
              maxItems: 8
              items: {oneOf: [{type: string, maxLength: 120}, {type: boolean}, {type: number}]}
    ServiceMetric:
      type: object
      required: [key, label, kind, unit, direction, value, status, warning, critical]
      properties:
        key: {type: string}
        label: {type: string}
        kind: {type: string, enum: [NUMBER, TEXT, BOOLEAN]}
        unit: {type: string}
        direction: {type: string, enum: [LOW, HIGH]}
        value: {oneOf: [{type: number}, {type: string}, {type: boolean}, {type: 'null'}]}
        status: {$ref: '#/components/schemas/Status'}
        warning: {type: [number, 'null']}
        critical: {type: [number, 'null']}
    ServiceInfo:
      type: object
      required: [title, description, dashboard, status, collected, metrics]
      properties:
        title: {type: string}
        description: {type: string}
        dashboard: {type: boolean}
        status: {$ref: '#/components/schemas/Status'}
        collected: {type: integer, minimum: 0}
        metrics: {type: array, maxItems: 16, items: {$ref: '#/components/schemas/ServiceMetric'}}
...
    Check:
      type: object
      required: [project_id, instance_id, check_id, name, category, direction, status,
        duration_ms, result_code, message, checked_at, history]
      properties:
        project_id: {$ref: '#/components/schemas/Identifier'}
        project_name: {type: [string, 'null']}
        instance_id: {$ref: '#/components/schemas/Identifier'}
        instance_name: {type: [string, 'null']}
        check_id: {$ref: '#/components/schemas/Identifier'}
        name: {type: string}
        category: {type: string, enum: [INTERNAL, API]}
        monitoring_enabled: {type: boolean, description: Central usage preference on this instance. Independent of agent discovery.}
        automatic_enabled: {type: boolean, description: Automatic execution preference. Manual execution is allowed when the check is used.}
        check_interval_seconds: {type: integer, minimum: 60, description: Effective automatic interval. API defaults are external 24 hours or internal 5 minutes; INTERNAL checks default to 5 minutes.}
        automatic_allowed: {type: boolean, description: Project and instance enabled; API checks also require the instance API automatic switch.}
        direction:
          oneOf:
            - {type: string, enum: [INTERNAL, EXTERNAL]}
            - {type: 'null'}
        status: {$ref: '#/components/schemas/Status'}
        duration_ms: {type: [integer, 'null']}
        result_code: {type: [string, 'null']}
        message: {type: [string, 'null']}
        checked_at: {type: [string, 'null'], format: date-time}
        history: {type: array, items: {$ref: '#/components/schemas/CheckHistorySample'}}
        http:
          oneOf: [{$ref: '#/components/schemas/HttpCheckDetails'}, {type: 'null'}]
        service_profile:
          oneOf: [{$ref: '#/components/schemas/ServiceProfile'}, {type: 'null'}]
        service_info:
          oneOf: [{$ref: '#/components/schemas/ServiceInfo'}, {type: 'null'}]
        details:
          oneOf:
            - type: object
              maxProperties: 32
              additionalProperties:
                oneOf:
                  - {type: string, maxLength: 1024}
                  - {type: number}
                  - {type: boolean}
                  - {type: 'null'}
            - {type: 'null'}
    HttpCheckDetails:
      type: object
      additionalProperties: false
      properties:
        method: {type: string, enum: [GET, POST]}
        url: {type: string, maxLength: 8192}
        request_query: {$ref: '#/components/schemas/HttpEvidenceHeaders'}
        request_headers: {$ref: '#/components/schemas/HttpEvidenceHeaders'}
        request_body: {type: [string, 'null'], maxLength: 4096}
        request_body_truncated: {type: boolean}
        connect_timeout_ms: {type: integer, minimum: 1}
        read_timeout_ms: {type: integer, minimum: 1}
        expected: {type: string, maxLength: 2048}
        status_code: {type: [integer, 'null'], minimum: 100, maximum: 599}
        response_headers: {$ref: '#/components/schemas/HttpEvidenceHeaders'}
        response_body: {type: [string, 'null'], maxLength: 4096}
        response_body_truncated: {type: boolean}
        response_bytes: {type: [integer, 'null'], minimum: 0}
    HttpEvidenceHeaders:
      type: object
      maxProperties: 32
      additionalProperties: {type: string, maxLength: 1024}
    CheckHistorySample:
      type: object
      additionalProperties: false
      required: [checked_at, duration_ms, status, result_code]
      properties:
        checked_at: {type: string, format: date-time}
        duration_ms: {type: [integer, 'null'], format: int64, minimum: 0}
        status: {$ref: '#/components/schemas/Status'}
        result_code: {type: [string, 'null']}
    CheckPage:
      allOf:
        - {$ref: '#/components/schemas/PageMetadata'}
        - type: object
          required: [items]
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/Check'}}
...